shared_requirements.txt
msrest
anyio
azure-ai-ml
azure-ai-resources
azure-common
azure-mgmt-authorization
azure-mgmt-core
typing-extensions
azure-core
six
isodate
azure-appconfiguration
azure-core-tracing-opentelemetry
azure-keyvault-secrets
azure-monitor-opentelemetry-exporter
azureml-telemetry
cryptography
msrestazure
requests
psutil
opencensus
opencensus-ext-azure
opencensus-ext-threading
opencensus-ext-logging
opentelemetry-api
azure-eventhub
aiohttp
mlflow
mlflow-skinny
msal
msal-extensions
azure-keyvault-certificates
azure-keyvault-keys
pyyaml
marshmallow
jsonschema
tqdm
strictyaml
colorama
pyjwt
azure-storage-blob
azure-storage-file-share
azure-storage-file-datalake
pydash
azure-mgmt-resourcegraph
azure-mgmt-resource
fixedint
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-django
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-psycopg2
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-urllib
opentelemetry-instrumentation-urllib3
opentelemetry-resource-detector-azure
azure-nspkg
azure-ai-nspkg
azure-cognitiveservices-nspkg
azure-mgmt-nspkg
azure-mixedreality-authentication
azure-schemaregistry
websocket-client
avro
uamqp
yarl
azure-identity
openai
azure-ai-vision-face